requirements.txt
asgiref==3.11.1
da-vinci==0.5.0
Django
django-thumbnails==0.8.0
djangorestframework
shortuuid==1.0.13
sorl-thumbnail
sqlparse==0.5.5
tzdata==2025.3
